refactor: Move format -> sourcenode/format
